Move entire maps ingestion system into /maps/ingestion subdirectory
