+++ src/pages/maps/ingestion/@id/components/cell/interval-selection.ts
