+++ src/pages/maps/ingestion/@id/components/panels.ts
