+++ src/pages/maps/ingestion/@id/components/table-interface.ts
