+++ src/pages/maps/ingestion/@id/table.d.ts
