+++ src/pages/maps/ingestion/main.module.sass
@@ -5,3 +5,6 @@
   border-radius: 0.2em
   display: inline-block
   font-size: 0.8em
+
+body
+  margin: 0
