+++ src/styles/core.sass
@@ -1,4 +1,17 @@
-// TODO: this file is a placeholder for when we migrate core styles to SASS
-  
 @use "@blueprintjs/colors/lib/scss/colors.scss" as colors
   
+.float-right
+  float: right
+
+.stick-to-top
+  position: sticky
+  top: 0
+  z-index: 1
+
+.stick-to-bottom
+  position: sticky
+  bottom: 0
+  z-index: 1
+
+.padding
+  padding: 1em
